requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
hedera-sdk-python
werkzeug
